requirements.txt
boto3==1.34.84
python-docx==1.1.0
python-dotenv==1.0.1
pypdf==4.1.0
